requirements.txt
ood-metrics==1.1.0
timm==0.6.12
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.64.1

# TODO fix:
faiss-gpu
pip install git+https://github.com/openai/CLIP.git
scikit-learn 
